fix clippy warnings for CI
field_reassign_with_default, type_complexity, needless_borrow,
needless_range_loop, unnecessary_cast, doc_lazy_continuation,
doc_overindented_list_items, useless_vec.

+++ tests/disk_template_e2e.rs
@@ -159,7 +159,7 @@ fn scenario_btrfs_filesystem_visible_at_dev_vda(
     // x86_64/aarch64 64-bit Linux) avoids the host-side
     // sign-extension trap that motivates the `compile_error!` in
     // `disk_template.rs`.
-    let fs_type = buf.f_type as i64;
+    let fs_type = buf.f_type;
     if fs_type != BTRFS_SUPER_MAGIC {
         anyhow::bail!(
             "/mnt/disk0 has statfs.f_type=0x{fs_type:x}, expected \
